Moved lexing error to tokens.
