+++ mini-pl/src/lexer/tokens.rs
@@ -1,8 +1,11 @@
 use std::fmt;
+use std::num::ParseIntError;
+use std::string::FromUtf8Error;
+use std::str::Utf8Error;
 
 use num_bigint::BigInt;
 
-use parsco::Sym;
+use parsco::{Sym, FromErr};
 
 use Ident;
 
@@ -59,7 +62,7 @@ impl Tok {
 
 #[derive(Clone, Debug, PartialEq)]
 pub enum Token {
-    Error(super::LexError),
+    Error(LexError),
     Identifier(Ident),
     Literal(Literal),
     Punctuation(Punctuation),
@@ -113,4 +116,53 @@ pub enum Keyword {
     Str,
     Bool,
     Assert,
+}
+
+#[derive(Debug, PartialEq, Clone)]
+pub enum LexError {
+    HexadecimalLexError(HexadecimalLexError),
+    UnknownEscape(String),
+    Unknown,
+}
+
+impl FromErr<()> for LexError {
+    fn from(_: ()) -> Self {
+        LexError::Unknown
+    }
+}
+
+impl FromErr<::parsco::common::Void> for LexError {
+    fn from(_: ::parsco::common::Void) -> Self {
+        unreachable!("Void is never type.")
+    }
+}
+
+impl FromErr<LexError> for LexError {
+    fn from(l: LexError) -> Self {
+        l
+    }
+}
+
+impl From<HexadecimalLexError> for LexError {
+    fn from(e: HexadecimalLexError) -> Self {
+        LexError::HexadecimalLexError(e)
+    }
+}
+
+#[derive(Debug, PartialEq, Clone)]
+pub enum HexadecimalLexError {
+    ParseIntError(ParseIntError),
+    FromUtf8Error(Utf8Error),
+}
+
+impl From<ParseIntError> for HexadecimalLexError {
+    fn from(e: ParseIntError) -> Self {
+        HexadecimalLexError::ParseIntError(e)
+    }
+}
+
+impl From<FromUtf8Error> for HexadecimalLexError {
+    fn from(e: FromUtf8Error) -> Self {
+        HexadecimalLexError::FromUtf8Error(e.utf8_error())
+    }
 }
